Fix CI: rustfmt + clippy clean
- cargo fmt --all applied across the repo (line-wrapping +
  collapsed-if formatting).
- Clippy fixes:
  - src/dsp/beat_tracker.rs: drop redundant `|off, frac| on_onset(off, frac)`
    closure in process_block_multichannel; pass `on_onset` directly.
  - src/lib.rs: collapse nested `if onset_method != … { if set_method().is_ok()
    { … } }` into a single condition.
  - tests/{beat_tracker,dataset_synthetic,link_listener,resilience}.rs:
    drop redundant `(0.020 * SR as f32) as f32` (same-type cast).
  - tests/link_listener.rs: pass the new `latency_offset_micros: i64 = 0`
    to `LinkPublisher::publish_tempo()` (signature changed in 56b37ce).

`cargo fmt --all -- --check`, `cargo clippy --all-targets --all-features
-- -D warnings`, and `cargo test` all pass locally.

+++ tests/beat_tracker.rs
@@ -13,7 +13,7 @@ const SR: u32 = 44_100;
 fn make_clicks(total_samples: usize, click_samples: &[usize]) -> Vec<f32> {
     let mut buf = vec![0.0f32; total_samples];
     let click_len = (0.050 * SR as f32) as usize; // ~50 ms
-    let decay_tau = (0.020 * SR as f32) as f32; // 20 ms decay
+    let decay_tau = 0.020 * SR as f32; // 20 ms decay
     for &t in click_samples {
         for i in 0..click_len {
             let pos = t + i;
